IMPALA-13653: Create hooks for Calcite planner in Frontend
This commit creates hooks into the Impala planner that will
call into the Calcite planner.

There is one factory hook and 3 compilation hooks introduced to
allow this to happen:

- CompilerFactory :  This factory hook has methods to create the
implementations for the three interfaces described below. The idea is
to introduce a query option that allows the Calcite objects to be called
instead of the Impala planner. In this iteration, the only
implementation called is the Impala specific implementations.

- ParsedStatement: The implementation of this interface will parse
the SQL query into an AST in its constructor. The implementation will
be a wrapper for the AST which can be accessed through the "getTopLevelNode"
method. For the current planner, the "StatementBase" object will be
accessible via this method.

- AnalysisDriver: The implementation of this interface will analyze the
parsed AST via the "analyze" method. To make the code review as easy
as possible, the AnalyzeDriverImpl implementation has been placed inside
the AnalysisContext object.

- SingleNodePlannerIntf: The implementation of this interface will
return a PlanNode that can run on a single node via the
"createSingleNodePlan" method. Various other methods in this
interface allow access to data structures that are needed for the final
plan. These objects are related to the output Expr objects.

Change-Id: I978aa48b160984ee5d36244cd915940f838d141d
Reviewed-on: http://gerrit.cloudera.org:8080/22306
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Aman Sinha <amsinha@cloudera.com>