IMPALA-13662: Bump the ARM toolchain to support ARM builds for RHEL 9
Pick up a new binary build of the current toolchain version for ARM.

The toolchain version is identical, the only difference is that the new
build added binaries for Rocky/RHEL 9 to the already supported OS
versions, reaching the same level of Impala build support as
Rocky/RHEL 8.

Tested by building Impala for RHEL9 for Intel and ARM both on private
infrastructure.

Change-Id: I5fd2e8c3187cb7829de55d6739cf5d68a09a2ed3
Reviewed-on: http://gerrit.cloudera.org:8080/22323
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>