IMPALA-13689: Fix webserver tests with curl 7.76.1
On Rocky Linux 9.5 a few checks started to fail because error 503
was expected to cause non 0 return value but it was not treated
as error. The difference is likely caused by the newer curl version.
curl documentation seems unclear about the return value of auth
related status codes.

The fix is to check the specific http status code instead of curl's
return value.

Testing:
- webserver tests passed on Rocky Linux 9.5

Change-Id: I354aa87a1b6283aa617f0298861bd5e79d03efc7
Reviewed-on: http://gerrit.cloudera.org:8080/22379
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>