IMPALA-13716: Calcite Planner: TupleIsNullPredicate fix for analytic functions
There is some special logic to materialize the TupleIsNullPredicate
functions that are created by join nodes for outer joins for analytic
functions. This commit refactors some of the code in the current Impala
planner and materializes them with the Analytic RelNode.

An example query from the test framework that causes this issue is:

select avg(g) over (order by f) af3
from alltypestiny t1
      left outer join
        (select
           id as a, coalesce(bigint_col, 30) as f,
           bigint_col  as g
         from alltypestiny) t2
      on (t1.id = t2.a);

Change-Id: Iaec363c2fa93a1e21bf74a40e5399e21ddd9bd60
Reviewed-on: http://gerrit.cloudera.org:8080/22411
Reviewed-by: Aman Sinha <amsinha@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>