IMPALA-13727: Fix wrong regex for hostname in test_multiple_blocks_mt_dop
The test uses r'host=(\S+:[0-9]*)' to match the hosts of all the
fragment instances in query profile, i.e.
  host_list = re.findall(r'host=(\S+:[0-9]*)', result.runtime_profile)
This is able to find the instance name in lines like
  Instance d843c27e276bfa7a:a86450f600000006 (host=name:27002):
and get "name:27002" from it.

However, the runtime profile we get right after fetching the query
result and before closing the query might have a section of total time
after the fragment instance host, e.g.
  Instance d843c27e276bfa7a:a86450f600000006 (host=name:27002):(Total: 19.999ms, non-child: 0.000ns, % non-child: 0.00%)

The regex pattern will match "name:27002):Total:" in the string. If not
all the instances have this "Total" section, the instance names mismatch
and the test will fail. See the query profile attached in the JIRA as an
example. Not sure how this happens. Maybe it's not guaranteed that the
profile is ready immediately when the last result row is fetched. There
might be a delay for the coordinator to process the final update of the
profile from executors. I think it's ok to fix the test first.

This fixes the regex pattern to be r'\(host=([^:]+:[0-9]*)\)' so it can
exactly match the instance name.

Tests
 - Verified the fix locally

Change-Id: If70daa344403f2ae8617bf86e7896c2bbfd9e736
Reviewed-on: http://gerrit.cloudera.org:8080/22457
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>