IMPALA-13781: Fix "Median Diff %" for report_benchmark_results.py
The calculation that report_benchmark_results.py uses for
"Median Diff %" is using the wrong divisor. For an AB test,
it should be dividing by the base median time (i.e. A), but
this is actually dividing by the new median time (i.e. B).
That gives the wrong range and can result in "Median Diff %"
values of >100%. This fixes the calculation to divide by the
base median time (i.e. A).

Change-Id: I6f9da14bc7d5b6ef6a795b9024760f8cae77d72f
Reviewed-on: http://gerrit.cloudera.org:8080/22532
Reviewed-by: Riza Suminto <riza.suminto@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>