IMPALA-13770 (Addendum): Close expressions in IcebergMergeCasePlan
IMPALA-13770 added code to call Close() on
IcebergMergeCase::{output_exprs_,filter_conjuncts_}. However, these
expressions are created by IcebergMergeCasePlan, with pointers to the
expressions copied to possibly multiple IcebergMergeCase objects.
Therefore, although it does not cause errors in practice, it is better
to close the expressions in IcebergMergeCasePlan.

This change adds a Close() method to IcebergMergeCasePlan that closes
these expressions.

This patch also calls Close() on IcebergMergeSinkConfig::merge_action_
and IcebergMergeSink::merge_action_evaluator_, which were not closed
previously.

Change-Id: Iefa998dea173051702ef08c03b489178a17a653f
Reviewed-on: http://gerrit.cloudera.org:8080/22522
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>