IMPALA-13787: Compensate for inaccuracies of EC2 instances in runtime-profile-test
The resolution of the 'MonotonicStopwatch' is less on some machine
configurations. On some EC2 instances, this can be 1 to 10 milliseconds.

If events are recorded into the event sequence between this 'jiffy',
the relation between the timestamps in the runtime-profile-test do not
hold, as events will be recorded to occur at the same time.

To properly simulate the events, either the test can be adapted to
compensate for inaccuracies or provide higher sleep time between events.

As this does not pose problems to the consistency of the recorded within
the profile JSON, only the sleep timer has been increased for running
tests on such machines.

Change-Id: I8150b7e38e7e5e5294e2dc61c16c2b102032cca2
Reviewed-on: http://gerrit.cloudera.org:8080/22526
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>